타입 정의 파일 도메인 별로 분리
