+++ src/types/common.ts
@@ -1,0 +1,5 @@
+export type Obj = Record<string, any>
+
+export type WithId<T> = T & { id: number }
+
+export type Nullable<T> = T | null
