+++ src/types/image.ts
@@ -1,0 +1,8 @@
+import { Nullable } from 'types/common'
+
+export type ImageFiles = Nullable<File[]>
+
+export type PreviewImage = {
+  url: string
+  altText: string
+}
