+++ src/utils/getValueOfKeyFromQueryString.ts
@@ -1,4 +1,4 @@
-import { Nullable } from 'types'
+import { Nullable } from 'types/common'
 
 type Value = Nullable<string>
